Cast STA as integer in surface_type_check

+++ src/service/points/validation/defects.py
@@ -227,7 +227,7 @@ class RouteDefectsValidation(RoutePointEventsValidation):
         ).select(
             msg=pl.format(
                 "Tipe perkerasan pada STA {} {} tidak sama dengan data RNI",
-                pl.col(self._events._sta_col).truediv(self._events.sta_conversion),
+                pl.col(self._events._sta_col).truediv(self._events.sta_conversion).cast(pl.Int64),
                 pl.col(self._events._lane_code_col)
             )
         )
